Adding a base integration test class and moving some files around
